feat(formulas.sls): ensure temporary date-based branch is removed
Up to now, there's been a slight issue with doing direct pushes to repos
when there are no changes to be pushed.  For example, we currently need
`libtofs.jinja` to be updated across formulas but not all of the
formulas have this file; when running across all formulas, this will
cause the temporary date-based branch to be created but then not
removed.  This commit is a quick fix for that, where the starting branch
is tracked and then checked out again at the end, before removing the
temporary branch (if it is still available).