feat(formulas): add basic management for `Gemfile+lock` formulas
These formulas also show up on the security warnings for the
organisation; ideally, they should be managed like the rest of the
formulas but implementing a basic solution for the `Gemfile` and
`Gemfile.lock` files right now.  Also includes management of the
`commitlint` GitHub Action as well as the `CODEOWNERS` file.

* backupninja-formula
* charles-formula
* dbeaver-formula
* diaspora-formula
* kubernetes-formula
* nextcloud-formula
* rlang-formula
* rstudio-formula
* wireguard-formula
* zookeeper-formula