feat(gemfile+lock): use `ssf` customised `inspec` repo
This has been pinned to `4.22.22` for a long time now, only because of a
one-line change that would prevent the diff from being suppressed when
using `kitchen-inspec`.  This now provides the latest version, which is
`4.41.0` at the time of this commit.

Confirmed doing this in the last Formulas Working Group meeting.